Move ndcube arithmetic tests to own module.
